# openapi.yaml
openapi: 3.1.0
info:
  title: Vultiserver API
  description: |
    TSS (Threshold Signature Scheme) server for keygen, keysign, and reshare
    operations supporting ECDSA, EdDSA, and MLDSA via GG20 and DKLS protocols.

    Vultiserver runs an Echo HTTP server and an asynq worker for async TSS
    operations backed by Redis. Most mutating endpoints enqueue an async task
    and return immediately.
  version: 1.0.0
  contact:
    name: Vultisig
    url: https://github.com/vultisig/vultiserver
  license:
    name: MIT
    url: https://github.com/vultisig/vultiserver/blob/main/LICENSE
servers:
  - url: https://api.vultisig.com/router
    description: Production
  - url: http://localhost:8080
    description: Local development
paths:
...
  /getDerivedPublicKey:
    get:
      operationId: getDerivedPublicKey
      summary: Derive a public key
      description: |
        Derives a child public key from a parent public key using the given
        chain code and derivation path. Supports both ECDSA and EdDSA curves.
      tags: [Utility]
      parameters:
        - name: publicKey
          in: query
          required: true
          description: Base public key (hex-encoded)
          schema:
            type: string
        - name: hexChainCode
          in: query
          required: true
          description: Chain code in hexadecimal format
          schema:
            type: string
        - name: derivePath
          in: query
          required: true
          description: BIP-32 style derivation path (e.g. `m/44'/60'/0'/0/0`)
          schema:
            type: string
        - name: isEdDSA
          in: query
          required: false
          description: Use EdDSA curve instead of ECDSA. Defaults to false.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Derived public key
          content:
            application/json:
              schema:
                type: string
                description: Hex-encoded derived public key
        "400":
          $ref: "#/components/responses/BadRequest"
  /vault/create:
    post:
      operationId: createVault
      summary: Create a new vault (keygen)
      description: |
        Initiates a new vault key generation ceremony. Enqueues an async
        `key:generation` (GG20) or `key:generationDKLS` (DKLS) task with a
        7-minute timeout. The server participates as one party in the TSS
        protocol.
      tags: [Vault]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultCreateRequest"
      responses:
        "200":
          description: Vault creation initiated
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /vault/create/check:
    post:
      operationId: checkVaultCreation
      summary: Check vault creation status
      description: |
        Polls the status of an in-progress vault key generation ceremony.
        Returns "ongoing" while the ceremony is still running, or the resulting
        public keys once the ceremony completes. The server retains ceremony
        state until this endpoint is called and returns a completed result.
      tags: [Vault]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultCreateCheckRequest"
      responses:
        "200":
          description: Ceremony status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultCreateCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: No ceremony found for the given session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /vault/reshare:
    post:
      operationId: reshareVault
      summary: Reshare vault key shares
      description: |
        Redistributes vault shares among a new set of parties. Enqueues an
        async `key:reshare` (GG20) or `key:reshareDKLS` (DKLS) task with an
        8-minute timeout. Requires a minimum 6-character encryption password.
      tags: [Vault]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReshareRequest"
      responses:
        "204":
          description: Reshare task enqueued
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /vault/migrate:
    post:
      operationId: migrateVault
      summary: Migrate vault from GG20 to DKLS
      description: |
        Converts an existing GG20 vault to the DKLS protocol. Enqueues an
        async `key:migrate` task with a 7-minute timeout.
      tags: [Vault]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MigrationRequest"
      responses:
        "204":
          description: Migration task enqueued
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /vault/import:
    post:
      operationId: importVault
      summary: Import an existing vault
      description: |
        Imports an existing vault via key import. The lib_type is forced to
        `KeyImport` (2). Enqueues an async `key:import` task with a 7-minute
        timeout.
      tags: [Vault]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyImportRequest"
      responses:
        "204":
          description: Import task enqueued
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /vault/get/{publicKeyECDSA}:
    get:
      operationId: getVault
      summary: Retrieve vault metadata
      description: |
        Retrieves and decrypts a vault backup file by its ECDSA public key.
        The encryption password must be provided via the `x-password` header
        (base64-encoded or plaintext).
      tags: [Vault]
      parameters:
        - $ref: "#/components/parameters/publicKeyECDSA"
        - name: x-password
          in: header
          required: true
          description: |
            Encryption password for vault decryption. May be base64-encoded
            or sent as plaintext; the server attempts base64 decoding first.
          schema:
            type: string
      responses:
        "200":
          description: Vault metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultGetResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /vault/exist/{publicKeyECDSA}:
    get:
      operationId: existVault
      summary: Check if a vault exists
      description: |
        Checks whether a vault backup file exists for the given ECDSA public key.
      tags: [Vault]
      parameters:
        - $ref: "#/components/parameters/publicKeyECDSA"
      responses:
        "200":
          description: Vault exists
        "404":
          description: Vault not found
  /vault/sign:
    post:
      operationId: signMessages
      summary: Sign messages with vault key
      description: |
        Enqueues a signing operation for one or more messages. Returns a task
        ID that can be used to poll for results. Enqueues an async `key:sign`
        (GG20) or `key:signDKLS` (DKLS) task with a 2-minute timeout.
      tags: [Signing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeysignRequest"
      responses:
        "200":
          description: Signing task enqueued; returns the async task ID
          content:
            application/json:
              schema:
                type: string
                description: Asynq task ID for polling the signing result
                example: "asynq:task:abc123"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /vault/mldsa:
    post:
      operationId: createMldsaVault
      summary: Add ML-DSA key to existing vault
      description: |
        Adds a post-quantum ML-DSA (Dilithium) key to an existing vault.
        The vault must already exist in block storage. Enqueues an async
        `key:createMldsa` task with a 5-minute timeout.
      tags: [Vault]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMldsaRequest"
      responses:
        "200":
          description: ML-DSA key creation task enqueued
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /vault/batch/keygen:
    post:
      operationId: createVaultBatch
      summary: Batch vault creation
      description: |
        Initiates vault creation across multiple protocols in a single request.
        Enqueues an async `key:keygenBatch` task with a 7-minute timeout.
      tags: [Batch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchVaultRequest"
      responses:
        "204":
          description: Batch keygen task enqueued
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /vault/batch/reshare:
    post:
      operationId: reshareVaultBatch
      summary: Batch vault reshare
      description: |
        Initiates vault resharing across multiple protocols in a single request.
        Enqueues an async `key:reshareBatch` task with an 8-minute timeout.
      tags: [Batch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchReshareRequest"
      responses:
        "204":
          description: Batch reshare task enqueued
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /vault/batch/import:
    post:
      operationId: importVaultBatch
      summary: Batch vault import
      description: |
        Imports vaults across multiple protocols in a single request. Protocols
        must not contain duplicates. Enqueues an async `key:importBatch` task
        with a 7-minute timeout.
      tags: [Batch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchImportRequest"
      responses:
        "204":
          description: Batch import task enqueued
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  parameters:
    publicKeyECDSA:
      name: publicKeyECDSA
      in: path
      required: true
      description: 66-character hex-encoded ECDSA public key identifying the vault
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{66}$"
        minLength: 66
        maxLength: 66
  responses:
    BadRequest:
      description: Validation error or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded (5 req/s, 30 burst, 5-min window)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "fail to parse request, err: invalid session_id"
    LibType:
      type: integer
      description: TSS library/protocol type
      enum: [0, 1, 2]
      x-enum-varnames: [GG20, DKLS, KeyImport]
      x-enum-descriptions:
        - GG20 threshold signature protocol
        - DKLS threshold signature protocol
        - Key import mode
...
    VaultCreateRequest:
      type: object
      required:
        - session_id
        - hex_encryption_key
        - hex_chain_code
        - encryption_password
      properties:
        session_id:
          type: string
          format: uuid
          description: Unique session identifier (UUID)
        hex_encryption_key:
          type: string
          description: 32-byte hex-encoded encryption key
          pattern: "^[0-9a-fA-F]{64}$"
        hex_chain_code:
          type: string
          description: Hex-encoded chain code for key derivation
        local_party_id:
          type: string
          description: Identifier for the local TSS party (optional; server assigns one if omitted)
        encryption_password:
          type: string
          description: Password used to encrypt the vault backup
        lib_type:
          $ref: "#/components/schemas/LibType"
    VaultCreateCheckRequest:
      type: object
      required:
        - session_id
      properties:
        session_id:
          type: string
          format: uuid
          description: Session identifier from the original /vault/create request
    VaultCreateCheckResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [ongoing, complete]
          description: '"ongoing" while the ceremony is running, "complete" when finished'
        public_key_ecdsa:
          type: string
          description: Hex-encoded ECDSA public key (present when status is "complete")
        public_key_eddsa:
          type: string
          description: Hex-encoded EdDSA public key (present when status is "complete")
...